Person: popularity is Double, not Integer.

+++ src/main/java/com/uwetrottmann/tmdb2/entities/BasePerson.java
@@ -8,7 +8,7 @@ public class BasePerson {
     public Boolean adult;
     public Integer id;
     public String name;
-    public Integer popularity;
+    public Double popularity;
     public List<Media> known_for;
 
 }
